complete automatically generate inverted-index  table & search

+++ HW1/Search/execute.sh
@@ -3,13 +3,17 @@
 #hdfs dfs -rm -r /user/TA/CalculateAverage/Output/
 #hadoop jar CalculateAverage.jar calculateAverage.CalculateAverage /user/shared/CalculateAverage/Input /user/TA/CalculateAverage/Output
 #hdfs dfs -cat /user/TA/CalculateAverage/Output/part-*
+if [ $# -lt 1 ]; then
+  echo "Please input search word"
+  exit 1
+fi
 
 your_hadoop_output_directory=HW1/output_invertedindex
 #INPUT=HW1/input
 INPUT=$your_hadoop_output_directory
 your_hadoop_output_directory=HW1/output_retrieval
 hdfs dfs -rm -r ${your_hadoop_output_directory}
-hadoop jar CalculateAverage.jar calculateAverage.CalculateAverage $INPUT ${your_hadoop_output_directory} cousin
+hadoop jar CalculateAverage.jar calculateAverage.CalculateAverage $INPUT ${your_hadoop_output_directory} $1
 
 rm output/*
 # hdfs dfs -cat ${your_hadoop_output_directory}/part-*
